rename json rpc methods files to entrypoint
