+++ overlore/jsonrpc/methods/generate_discussion/entrypoint.py
